fix(schema): align PD DB schema inconsistencies (PRI-473) (#1066)
* fix(schema): align PD DB schema inconsistencies (score type, FK checks, orphan table, doc alignment)

- P0-1: pain_events.score INTEGER to REAL (SQLite type affinity fix)

- P0-2: DATA_ARCHITECTURE.md approvals fields aligned with production code

- P0-3: approval store uses UPDATE...RETURNING (eliminate race window, ERR-015)

- P1-3: application-layer FK checks in 3 stores (enqueue/createArtifact/recordActivation)

- P1-5: DATA_ARCHITECTURE.md implementation status annotations (post-MVP table markers)

- P3-12: drop orphaned confirm_first_state table DDL

Test coverage: 3 new FK validation test files (6 tests), 4 regression test files updated. 287 activation + 544 architecture tests pass. ERR checklist: ERR-001/005/002/009/010/015/004/008. Closes PRI-473 (batch 1+2)

* fix(schema): fix ERR-001 as-cast violations and add schema_version table (P2-10)

- ERR-001: replace 'as ApprovalRow | undefined' on RETURNING * results with isRecord() + mapRowToRecord() type guard pattern (mirrors sqlite-activation-state-store.ts)

- P2-10: add schema_version table to state.db with getSchemaVersion()/setSchemaVersion() methods; core cannot import plugin MigrationRunner so this is a lightweight version

- Tests: 5 new tests for schema_version lifecycle (create, default, increment, persist, no-dup)

Closes PRI-473

* test(pi-artifact): seed parent tasks for SQLite FK validation (P1-3)

* fix(runtime-v2): seed parent records for P1-3 FK validation in demo/approval/activation callers

The P1-3 application-layer FK checks added in 6c02af53 (createArtifact,
enqueue, recordActivation) reject references to non-existent parent rows.
Three call sites did not seed the parent row, breaking downstream tests:

- demo-story-a-runner: createArtifact(task-demo-${runId}) failed because the
  tasks row was never seeded. Seed it before artifact creation. Fixes
  pd-cli/demo-story-a (12) and create-principles-disciple/smoke (1).
- approvals-api.test: enqueue(artifact-proven-${ch}) failed because the
  pi_artifacts row was never seeded. Seed via seedPrincipleArtifact before
  enqueue. Fixes pd-console/approvals-api (1).
- runtime-v2-prompt-activation.test: the "missing activated artifact" test
  intentionally references a non-existent artifact. recordActivation now
  rejects it, so insert the dangling activation directly via DB (bypassing
  the store FK check), preserving the test's graceful-degradation intent.
  Fixes openclaw-plugin unit + coverage (2).

ERR entries considered: ERR-009/ERR-010 (fail loud on missing required
field), ERR-002 (graceful degradation with reason). The dangling-activation
test bypasses the store FK check deliberately to simulate legacy data; the
hook still fails loud with a logged warning, satisfying ERR-002.

+++ packages/principles-core/src/runtime-v2/__tests__/pain-signal-observability.test.ts
@@ -114,7 +114,7 @@ describe('recordPainSignalObservability', () => {
         id INTEGER PRIMARY KEY AUTOINCREMENT,
         session_id TEXT NOT NULL,
         source TEXT NOT NULL,
-        score INTEGER NOT NULL,
+        score REAL NOT NULL,
         reason TEXT,
         severity TEXT,
         origin TEXT,
